requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
astroid==3.3.8
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
dill==0.3.9
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
packaging==24.2
platformdirs==4.3.6
playwright==1.49.1
propcache==0.2.1
pyee==12.0.0
pylint==3.3.3
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
TikTokApi==6.5.2
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
yt-dlp==2024.12.23
